vcf/reader: Move lazy record reader to a module
